requirements.txt
streamlit
pandas
numpy
yfinance
plotly
matplotlib
seaborn
scikit-learn
tensorflow
transformers
torch
requests
python-dotenv